% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{extract_groupings}
\alias{extract_groupings}
\title{Extract column indexes and names from a formula}
\usage{
extract_groupings(formula, mm, has_re = FALSE)
}
\arguments{
\item{formula}{formula used to create the model matrix}

\item{mm}{model matrix (either from lme4:::lFormula or 
MatrixModels::model.Matrix}

\item{has_re}{whether the model includes random effects}
}
\value{
list with the column indexes for each term, the
        variables involved in each term, and a vector 
        with on (term) name for each column.  This does
        not (yet) extract the terms on the LHS of the
        random effects formulas separately.
}
\description{
Extract column indexes and names from a formula
}
