% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_re_coefficient_names}
\alias{extract_re_coefficient_names}
\title{extract_re_coefficient_names
(migrated from 02-setup.functions.R)}
\usage{
extract_re_coefficient_names(fmm)
}
\description{
extract_re_coefficient_names
(migrated from 02-setup.functions.R)
}
