% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{extract_response}
\alias{extract_response}
\title{Extract the formula response column form a data frame}
\usage{
extract_response(formula, data)
}
\arguments{
\item{formula, }{a formula}

\item{data}{data frame to fetch from}
}
\value{
vector of the response column
}
\description{
Extract the formula response column form a data frame
}
