% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{extract_response_name}
\alias{extract_response_name}
\title{Extract response name}
\usage{
extract_response_name(formula)
}
\arguments{
\item{formula}{formula to parse}
}
\value{
name of LHS variable.
}
\description{
Extract response name
}
