% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{flat_mm}
\alias{flat_mm}
\title{Create a sparse list model matrix}
\usage{
flat_mm(formula = 1, data = NULL, ...)
}
\arguments{
\item{formula}{formula to use}

\item{data}{where to look up terms}
}
\value{
list with Stan-friendly components
}
\description{
Create a sparse list model matrix
}
