% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{flip_nze}
\alias{flip_nze}
\title{Turn a row-wize nze into a col-wise NZE and vice-versa.}
\usage{
flip_nze(nze, n)
}
\description{
Turn a row-wize nze into a col-wise NZE and vice-versa.
}
