% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-formulate.R
\name{formulate}
\alias{formulate}
\title{Constructs a list containing the model matrix for the formulas,
data, and configurations provided.}
\usage{
formulate(
  models = list(),
  data,
  configuration = list(),
  auxiliary_data = list()
)
}
\arguments{
\item{models}{a list of \code{formula}s. Each formula creates its
own model matrix.}

\item{data}{\code{data.frame} to pull formula terms from.}

\item{configuration}{configuration}

\item{auxiliary_data}{auxiliary data}
}
\value{
List representing the model matrix constructed from
   the parameters. The named list contains:
   1. \code{models}: the list of models supplied
   2. \code{data}: the data used to evaluate the terms of the model matrix
   3. \code{configuration}: any additional configuration data used
   4. \code{inputs}: the sparse-format model matrix as stan-formatted input
   5. \code{matrices}: the list-format model matrix for each formula
   6. \code{auxiliary_data}: any auxiliary data used
}
\description{
\code{formulate} constructs the model matrix as a list-format
model matrix per formula, the data used to evaluate the
terms for the model matrix, the model matrix as a sparse
matrix in a container object, additioanl configuration data
used, and the list of models supplied.
}
