% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{get_id_idx}
\alias{get_id_idx}
\title{For a code return the index if a code exists or a zero otherwise}
\usage{
get_id_idx(code, table)
}
\arguments{
\item{code}{}

\item{table}{h_table object.}
}
\description{
For a code return the index if a code exists or a zero otherwise
}
