% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{get_leaf_idxs}
\alias{get_leaf_idxs}
\title{Get the leaf indexes for an 'h_table'}
\usage{
get_leaf_idxs(table)
}
\arguments{
\item{table}{'h_table'}
}
\description{
A leaf index is one that does _not_ appear as a parent.
}
