% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{get_nze}
\alias{get_nze}
\title{Get non-zero entries of a matrix in a list of one item per row/column}
\usage{
get_nze(m, compression = "row")
}
\description{
Get non-zero entries of a matrix in a list of one item per row/column
}
