% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{get_zero_idx}
\alias{get_zero_idx}
\title{Return the index of the zero-value entry in the parameter index}
\usage{
get_zero_idx()
}
\value{
index of the zero entry in the 'h_table'
}
\description{
Return the index of the zero-value entry in the parameter index
}
