% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\docType{package}
\name{hierarchy}
\alias{hierarchy}
\title{hierarchy: A package for creating sensible hierarchical data structures!}
\description{
The hierarchy package provides ways to take flat data from csv files to
a model matrix that has the right data structure for use in hierarchical models.
}
