\name{id_to_levels}
\alias{id_to_levels}
\title{Path enum to levels}
\usage{
  id_to_levels(ids, sep = ".")
}
\arguments{
  \item{ids}{path enum vector}

  \item{sep}{id seperator (e.g. ".")}
}
\description{
  Convert a path enumeration id vector to a vector
  containing hierarchical levels
}
\examples{
# id_to_levels(notes$Id)
}

