% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{imbue}
\alias{imbue}
\title{Interpret a formula in the context of some data.}
\usage{
imbue(x, e, s)
}
\arguments{
\item{x}{formula}

\item{e}{environment to use for data, can be a list}

\item{s}{names of state variables to be parsed but not
included.}
}
\value{
list with all formula terms defined from the 
        environment.
}
\description{
Interpret a formula in the context of some data.
}
