% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{involves}
\alias{involves}
\title{Calculate which variables are involved in which 
simple or interaction term in the formula}
\usage{
involves(x)
}
\arguments{
\item{x}{formula}
}
\value{
list, one entry per term, one character vector
              per entry listing the involved data items
              states, or calls.
}
\description{
Calculate which variables are involved in which 
simple or interaction term in the formula
}
