% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{is_interaction}
\alias{is_interaction}
\title{Identify interactions.}
\usage{
is_interaction(x)
}
\arguments{
\item{x}{call to check}
}
\value{
TRUE iff the call is an interaction.
}
\description{
A call is an interaction if it is a simple call 
(?is_simple) that contains no '+' operators.  This
is significant because it can be processed as a single
contrast/covariate.
}
