% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{is_simple}
\alias{is_simple}
\title{Identify 'simple' call componenets.}
\usage{
is_simple(x, simple = c("+", ":", "("))
}
\arguments{
\item{x}{call to check}

\item{simple}{vector of operators to accept as simple.}
}
\value{
TRUE iff the call is 'simple'.
}
\description{
A call is 'simple' if it only includes the prescribed
operators in addition to arbitrary unary functions.
This is significant because such a call describes a 
linear combination without further parsing.
}
