% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{just_names}
\alias{just_names}
\title{Pull out just names from a call.}
\usage{
just_names(node)
}
\arguments{
\item{node}{a call object.}
}
\value{
pairlist of names
}
\description{
Pull out just names from a call.
}
