% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-csv.R
\name{load_csv}
\alias{load_csv}
\title{Load a .csv file in a round-trip-friendly way}
\usage{
load_csv(file, .table_name = NULL)
}
\arguments{
\item{file}{path to .csv file with read access.}

\item{table_name}{}
}
\value{
tibble of file contents.
}
\description{
Load a .csv file in a round-trip-friendly way
}
