% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{m_as_list}
\alias{m_as_list}
\title{Process fixed effects model}
\usage{
m_as_list(m)
}
\arguments{
\item{MatrixModels::model.Matrix}{object}
}
\value{
list with model matrix components and labels
}
\description{
Process fixed effects model
}
