% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{np_rows}
\alias{np_rows}
\title{Make h_table no-parent rows}
\usage{
np_rows(table)
}
\arguments{
\item{table}{h_table to check for no-parent codes.}
}
\value{
data frame rows for all no-parent rows (with ZERO as parent).
}
\description{
The no-parent rows are the parents of all regions that
themselves have no parent regions.  These are identified
as the rows who's parent_region_id does not appear in the 
'id' column.  They are added so that all parts of the hierarchy
terminate at a well-defined "parent_region_level" regardless of the
depth of the hierarchy.
}
