% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{parse_re_terms}
\alias{parse_re_terms}
\title{Basic parse to get the LHS and RHS of an lme4 random effect formula.}
\usage{
parse_re_terms(name)
}
\arguments{
\item{name}{term within parents in formula, e.g., (1 | a).}
}
\value{
names of lhs and rhs terms, e.g., c(lhs = 1, rhs = "a")
}
\description{
Basic parse to get the LHS and RHS of an lme4 random effect formula.
}
