% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{region_names}
\alias{region_names}
\title{Get names of levels from a hierarchy table.}
\usage{
region_names(table)
}
\arguments{
\item{table}{table generated by `spatial_hierarchy`}
}
\value{
names names of hierarchy levels
}
\description{
Get names of levels from a hierarchy table.
}
