% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-formulate.R
\name{setup_configuration}
\alias{setup_configuration}
\title{Setup configuration list for \code{formulate}.}
\usage{
setup_configuration(configuration, response_names)
}
\arguments{
\item{configuration}{named \code{list}}

\item{response_names}{\code{vector} of characters that need
to be found in the data frame.}
}
\value{
named \code{list} where each response_name either
   contains a configuration that's not NULL or an empty list.
}
\description{
For each name in \code{response_names}, it will either
find the configuration inside the list or it will put
an empty list there.
}
