% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-formulate.R
\name{setup_data}
\alias{setup_data}
\title{Setup data for \code{formulate}.}
\usage{
setup_data(data, response_names)
}
\arguments{
\item{data}{\code{data.frame} with at least one row of data}

\item{response_names}{\code{vector} of characters that need
to be found in the data frame.}
}
\value{
\code{data.frame} where each \code{response_name} is
   found as a column
}
\description{
For each name in \code{response_names}, it will either
find a column with that name or add a new column filled
with 1s.
}
