% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{simplify}
\alias{simplify}
\title{Simplify formulas to a standard form using}
\usage{
simplify(node, side = rhs, debug = FALSE)
}
\arguments{
\item{node}{an R formula object.}

\item{side}{which side of the formula to parse, either the `rhs` (default)
or the `lhs` functions.}

\item{debug}{iff TRUE we print the current node as we parse.}
}
\value{
a simplified formula involving _only_ undefined functions, 
          intercept-related calls, parens, '+', and ':'
}
\description{
Applies a recursive descent parser to transform
'*', grouping using parentheses, and lme4-style 
random effects.  `1` is for an intercept, 
`0` is for an explicit no-intercept.  Processing
of contrasts is left for later.
}
