% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{spatial_hierarchy}
\alias{spatial_hierarchy}
\title{Generate a single hierarchy table for an implied hierarchy}
\usage{
spatial_hierarchy(files)
}
\arguments{
\item{files}{paths to .csv files describing spatial relationships}
}
\value{
single hierarchy table
}
\description{
The first item describes the terminal level for all levels.
Each item describes one level of the hierarchy.  That's an 'h_table'.
}
