% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-model-matrix-helpers.R
\name{st}
\alias{st}
\title{Inefficiently transpose a sparse matrix}
\usage{
st(m)
}
\arguments{
\item{m}{sparse matrix}
}
\value{
transpose of m
}
\description{
Inefficiently transpose a sparse matrix
}
