% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{subterms}
\alias{subterms}
\title{Simplify interaction terms.}
\usage{
subterms(node)
}
\arguments{
\item{a}{std form formula}
}
\value{
formula with labelled calls to 'term' function.
}
\description{
Also label what are 'term' objects
}
