% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{term_list}
\alias{term_list}
\title{Terms as list}
\usage{
term_list(node)
}
\arguments{
\item{node}{std-form formula}
}
\value{
list with lists of calls for lhs/rhs of formula
}
\description{
Calls from a std-form formula
}
