% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{zero_row}
\alias{zero_row}
\title{Make h_table zero row.}
\usage{
zero_row()
}
\value{
an 'h_table' zero_Row
}
\description{
This row is the parent of the parents of rows with no parents.
It exists so that all rows have parents, even the rows that are
constructed to be the parents of rows with no parents...
}
