% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{expand}
\alias{expand}
\title{Expand factors into model matrix blocks}
\usage{
expand(x, methods = hierarchy:::default_expand_methods())
}
\arguments{
\item{x}{result of `imbue`}
}
\value{
per-term list of lists of matrices
}
\description{
Expand factors into model matrix blocks
}
