% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-formulate.R
\name{formulate}
\alias{formulate}
\title{Create setting for a model matrix}
\usage{
formulate(models = list(), data)
}
\arguments{
\item{models, }{a list of formulas.  One formula per}

\item{data}{data.frame to pull formula terms from.}
}
\value{
list containing 1) the list-format model
        matrix for each formula; 2) the data used 
        to evaluate terms for the matrix model; 
        and 3) the sparse-format model matrix in a
        container object.
}
\description{
Calling this function constructs the (sparse) model
matrices for all models and combines the data
into a single list with all necessary prefixes.
}
