% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\docType{class}
\name{hierarchy-class}
\alias{hierarchy-class}
\alias{hierarchy_factory}
\title{A class for indexing a hierarchy in a model}
\description{
The class holds hierarchical relationships over all timepoints, and can
return indexes for any given level and all parents.
}
\section{Fields}{

\describe{
\item{\code{.levels}}{names of the (child) levels in the hierarchy}

\item{\code{.table}}{table representing the hierarchy}
}}

\section{Methods}{

\describe{
\item{\code{get_size(timepoint = NULL)}}{Get the number of parameters in the _entire_ hierarchy at a given timepoint, including the zero parameter.}

\item{\code{get_table(timepoint = NULL)}}{Method to retrieve the state of the hierarchy at a given point in time (date)}

\item{\code{initialize(files, ...)}}{Construct hierarchy from given relationship files.}
}}

