% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{imbue}
\alias{imbue}
\title{Interpret a formula in the context of some data.}
\usage{
imbue(terms, data, methods = hierarchy:::default_imbue_methods())
}
\arguments{
\item{t}{LHS or RHS of 'terms' calculated by 'involves'}

\item{e}{environment to use for data, can be a list}
}
\value{
list with all formula terms defined from the 
        environment.
}
\description{
Interpret a formula in the context of some data.
}
