% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{load_spatial_relationship}
\alias{load_spatial_relationship}
\title{Load relationship file}
\usage{
load_spatial_relationship(file)
}
\arguments{
\item{file}{path to .csv file describing a spatial relationship}
}
\value{
table describing the spatial relationship
}
\description{
Load relationship file
}
