% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-helpers.R
\name{np_rows}
\alias{np_rows}
\title{Make h_table no-parent rows}
\usage{
np_rows(table)
}
\arguments{
\item{table}{h_table to check for no-parent codes.}
}
\value{
data frame rows for all no-parent rows (with ZERO as parent).
}
\description{
Make h_table no-parent rows
}
